Move RestApiDescription to separate component
