+++ Source/SelfService/Web/integrations/connection/connectionDetails/consumeDataRestAPI/index.tsx
@@ -6,9 +6,9 @@ import React, { useReducer } from 'react';
 import { useSnackbar } from 'notistack';
 import { useQueryClient } from '@tanstack/react-query';
 
-import { Box, Typography, Select, MenuItem } from '@mui/material';
+import { Select, MenuItem } from '@mui/material';
 
-import { Button, ContentContainer, ContentHeader, ContentParagraph, ContentSection, IconButton, Link, StatusIndicator, StatusIndicatorProps, Switch } from '@dolittle/design-system';
+import { Button, ContentContainer, ContentHeader, ContentParagraph, StatusIndicatorProps, Switch } from '@dolittle/design-system';
 import { RestApiServiceStatus } from '../../../../apis/integrations/generated';
 import { useConnectionsIdRestApiStatusGet, useConnectionsIdRestApiEnablePost, useConnectionsIdRestApiDisablePost } from '../../../../apis/integrations/connectionRestApiApi.hooks';
 import { CACHE_KEYS } from '../../../../apis/integrations/CacheKeys';
@@ -16,6 +16,7 @@ import { useConnectionIdFromRoute } from '../../../routes.hooks';
 import { CredentialsContainer } from './Credentials/CredentialsContainer';
 import { EnableRestApiSection } from './EnableRestApiSection';
 import { DisableRestApiDialog, disableRestApiDialogReducer } from './DisableRestApiDialog';
+import { RestApiDescriptionSection } from './RestApiDescriptionSection';
 
 /* The ERP ReadModels -service must be specific to the connection, so we need
     to generate the URL dynamically. */
@@ -39,21 +40,9 @@ export const ConsumeDataRestAPIView = () => {
     const [disableServiceDialogState, disableServiceDialogDispatch] = useReducer(disableRestApiDialogReducer, { isOpen: false });
 
 
-    const handleRestApiLinkCopy = () => {
-        navigator.clipboard.writeText(restApiUrl);
-        enqueueSnackbar('Rest API URL copied to clipboard.');
-    };
-
-    const handleOpenApiDocumentationLinkCopy = () => {
-        navigator.clipboard.writeText(openApiDocumentationUrl);
-        enqueueSnackbar('OpenAPI documentation copied to clipboard.');
-    };
-
     const serviceStatus = forcedServiceStatus || apiStatus?.service;
     const showEnableSection = (apiStatus?.target === 'Disabled') || serviceStatus === 'Deploying' || forceShowEnable;
 
-    const restApiUrl = `${apiStatus?.basePath}swagger/index.html`;
-    const openApiDocumentationUrl = `${apiStatus?.basePath}swagger/v1/swagger.json`;
 
     const statusIndicatorFromServiceStatus = (status: RestApiServiceStatus | ''): StatusIndicatorProps['status'] => {
         switch (status?.toLocaleLowerCase()) {
@@ -151,41 +140,7 @@ export const ConsumeDataRestAPIView = () => {
                         status={serviceStatus || 'Off'}
                         isEnabling={enableMutation.isLoading}
                     />
-                    : <>
-                        <ContentSection title='Rest API URL'>
-                            <Box sx={{ display: 'flex', alignItems: 'center', pt: 2, gap: 1 }}>
-                                <Link
-                                    target
-                                    href={restApiUrl}
-                                    message={restApiUrl}
-                                />
-                                <IconButton
-                                    tooltipText='Copy rest API URL link to clipboard'
-                                    icon='CopyAllRounded'
-                                    color='primary'
-                                    onClick={handleRestApiLinkCopy}
-                                />
-                            </Box>
-                        </ContentSection>
-
-                        <ContentSection title='Rest API Documentation'>
-                            <Typography sx={{ pt: 1.5 }}>Our rest API is documented using OpenAPI.</Typography>
-                            <Box sx={{ display: 'flex', alignItems: 'center', pt: 2, gap: 1 }}>
-                                <Link
-                                    target
-                                    ariaLabel='OpenAPI documentation'
-                                    href={openApiDocumentationUrl}
-                                    message={openApiDocumentationUrl}
-                                />
-                                <IconButton
-                                    tooltipText='Copy OpenAPI documentation link to clipboard'
-                                    icon='CopyAllRounded'
-                                    color='primary'
-                                    onClick={handleOpenApiDocumentationLinkCopy}
-                                />
-                            </Box>
-                        </ContentSection>
-                    </>
+                    : <RestApiDescriptionSection restApiBaseUrl={apiStatus?.basePath || ''} />
                 }
                 <CredentialsContainer />
             </ContentContainer>
